Move data package into the components package
COSDK-500
