+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/remote/DefaultAnalyticsSetupProvider.kt
@@ -14,7 +14,7 @@ import com.adyen.checkout.core.analytics.internal.AnalyticsParamsLevel
 import com.adyen.checkout.core.analytics.internal.AnalyticsPlatformParams
 import com.adyen.checkout.core.analytics.internal.AnalyticsSource
 import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsSetupRequest
-import com.adyen.checkout.core.data.model.Amount
+import com.adyen.checkout.core.components.data.model.Amount
 import java.util.Locale
 
 @Suppress("LongParameterList")
