+++ core/src/main/java/com/adyen/checkout/core/components/CheckoutConfiguration.kt
@@ -17,7 +17,7 @@ import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.analytics.AnalyticsConfiguration
 import com.adyen.checkout.core.common.internal.helper.CheckoutConfigurationMarker
 import com.adyen.checkout.core.common.internal.helper.LocaleUtil
-import com.adyen.checkout.core.data.model.Amount
+import com.adyen.checkout.core.components.data.model.Amount
 import com.adyen.checkout.core.internal.Configuration
 import kotlinx.parcelize.IgnoredOnParcel
 import java.util.Locale
