+++ core/src/main/java/com/adyen/checkout/core/internal/BaseConfigurationBuilder.kt
@@ -12,7 +12,7 @@ import android.content.Context
 import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.analytics.AnalyticsConfiguration
 import com.adyen.checkout.core.common.internal.helper.LocaleUtil
-import com.adyen.checkout.core.data.model.Amount
+import com.adyen.checkout.core.components.data.model.Amount
 import java.util.Locale
 
 @Suppress(
