+++ core/src/main/java/com/adyen/checkout/core/sessions/internal/data/api/SessionRepository.kt
@@ -10,8 +10,8 @@ package com.adyen.checkout.core.sessions.internal.data.api
 
 import androidx.annotation.RestrictTo
 import com.adyen.checkout.core.common.internal.helper.runSuspendCatching
+import com.adyen.checkout.core.components.data.PaymentComponentData
 import com.adyen.checkout.core.components.paymentmethod.PaymentMethodDetails
-import com.adyen.checkout.core.data.PaymentComponentData
 import com.adyen.checkout.core.sessions.SessionModel
 import com.adyen.checkout.core.sessions.internal.data.model.SessionPaymentsRequest
 import com.adyen.checkout.core.sessions.internal.data.model.SessionPaymentsResponse
