+++ core/src/main/java/com/adyen/checkout/core/sessions/internal/data/model/SessionPaymentsResponse.kt
@@ -9,7 +9,7 @@
 package com.adyen.checkout.core.sessions.internal.data.model
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.data.OrderResponse
+import com.adyen.checkout.core.components.data.OrderResponse
 import com.adyen.checkout.core.exception.ModelSerializationException
 import com.adyen.checkout.core.internal.data.model.ModelObject
 import com.adyen.checkout.core.internal.data.model.ModelUtils
