+++ core/src/main/java/com/adyen/checkout/core/sessions/internal/data/model/SessionSetupResponse.kt
@@ -9,8 +9,8 @@
 package com.adyen.checkout.core.sessions.internal.data.model
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.data.model.Amount
-import com.adyen.checkout.core.data.model.PaymentMethodsApiResponse
+import com.adyen.checkout.core.components.data.model.Amount
+import com.adyen.checkout.core.components.data.model.PaymentMethodsApiResponse
 import com.adyen.checkout.core.exception.ModelSerializationException
 import com.adyen.checkout.core.internal.data.model.ModelObject
 import com.adyen.checkout.core.internal.data.model.ModelUtils
@@ -56,13 +56,13 @@ data class SessionSetupResponse(
                         PAYMENT_METHODS,
                         ModelUtils.serializeOpt(
                             modelObject.paymentMethodsApiResponse,
-                            PaymentMethodsApiResponse.SERIALIZER
-                        )
+                            PaymentMethodsApiResponse.SERIALIZER,
+                        ),
                     )
                     jsonObject.putOpt(RETURN_URL, modelObject.returnUrl)
                     jsonObject.putOpt(
                         CONFIGURATION,
-                        ModelUtils.serializeOpt(modelObject.configuration, SessionSetupConfiguration.SERIALIZER)
+                        ModelUtils.serializeOpt(modelObject.configuration, SessionSetupConfiguration.SERIALIZER),
                     )
                     jsonObject.putOpt(SHOPPER_LOCALE, modelObject.shopperLocale)
                 } catch (e: JSONException) {
@@ -80,12 +80,12 @@ data class SessionSetupResponse(
                         expiresAt = jsonObject.getStringOrNull(EXPIRES_AT).orEmpty(),
                         paymentMethodsApiResponse = ModelUtils.deserializeOpt(
                             jsonObject.optJSONObject(PAYMENT_METHODS),
-                            PaymentMethodsApiResponse.SERIALIZER
+                            PaymentMethodsApiResponse.SERIALIZER,
                         ),
                         returnUrl = jsonObject.getStringOrNull(RETURN_URL),
                         configuration = ModelUtils.deserializeOpt(
                             jsonObject.optJSONObject(CONFIGURATION),
-                            SessionSetupConfiguration.SERIALIZER
+                            SessionSetupConfiguration.SERIALIZER,
                         ),
                         shopperLocale = jsonObject.getStringOrNull(SHOPPER_LOCALE),
                     )
