+++ core/src/test/java/com/adyen/checkout/core/analytics/internal/data/remote/DefaultAnalyticsSetupProviderTest.kt
@@ -16,7 +16,7 @@ import com.adyen.checkout.core.analytics.internal.AnalyticsParamsLevel
 import com.adyen.checkout.core.analytics.internal.AnalyticsPlatformParams
 import com.adyen.checkout.core.analytics.internal.AnalyticsSource
 import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsSetupRequest
-import com.adyen.checkout.core.data.model.Amount
+import com.adyen.checkout.core.components.data.model.Amount
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.mockito.kotlin.doReturn
